toy around with doing things via everything included into one project, but this... actually seems perhaps less sane, because i don't really know how to control scope this way; invoking many ants seems cray, but apparently can work quickly in this situation, and is more practical to control than trying to wrap leagues of stuff in giant macrodefs.
on the other hand... having a the root build pass much information to the sub-builds mean those break when run alone, and that's not cool either.  maybe the solution is to have a bunch of defaults set in a build-common config, and then... yarhg, i dunno.

i also don't realllly want the dependencies file to be in the root, but... if the root interacts with sub-builds by ant invocation, the sub-builds can't really pass data back up.  hm, i suppose i can keep splitting subfiles out and expect those to be loaded by either side.